release(2.1.0): experimental voxel_graph feature
Voxel graph allows you to accomodate self-touch shapes by using an
image of bitfields that describe the permissible directions of travel
at each voxel. The distance from a voxel to an impermissible face is
treated as 0.5 while the distance to a neighboring black voxel is
treated as 1.0.

The little endian 8 bit field is specified as:

_ _ -z +z -y +y -x +x

The first two bits are not used. See the connected-components-3d library
for a helper function for generating bitfield images.